# docs/api.yml
openapi: 3.1.0
info:
  title: Food Bank Finder & Management API
  version: 1.0.0
  description: jade-jasmine team api
paths:
...
  /foodbank:
    get:
      tags:
        - completed
        - food bank
      summary: Get food bank list
      security: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - name
                    - street
                    - city
                    - country
                    - secondary_admin_id
                    - charity_id
                    - hours
                    - admin
                  properties:
                    id:
                      type: integer
                    fax:
                      type: integer
                    city:
                      type: string
                    name:
                      type: string
                    admin:
                      $ref: "#/components/schemas/User"
                    hours:
                      type: string
                    phone:
                      type: integer
                    staff:
                      type: array
                      items:
                        type: integer
                        minimum: 1
                        description: the id of the employee
                    street:
                      type: string
                      examples:
                        - 555 Food Bank Way
                        - 55 Pantry St
                    country:
                      type: string
                    unit_no:
                      type: string
                      description: Can be a number or a letter
                    website:
                      type: string
                    latitude:
                      type: number
                      format: float
                    province:
                      type: string
                    longitude:
                      type: number
                      format: float
                    charity_id:
                      type: string
                    postal_code:
                      type: string
                    secondary_admin_id:
                      type: string
                  description: The food bank or food pantry entity and all its details.
                title: GetFoodBankListOk
          description: The request was successful, and the server has returned the
            requested resource in the response body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - $ref: "#/components/parameters/city"
        - $ref: "#/components/parameters/fb_name"
        - in: query
          name: fb_id
          style: form
          schema:
            type: integer
          explode: true
          required: false
          description: the id of the food bank
          allowReserved: false
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      description: >-
        Fetches a comprehensive list of all food banks in our db. Each food bank
        object includes essential details but doesn't include hours or staff.

        Also doesn't include the admin id as that is protected information.


        If the request is made by an authenticated user who also happens to be
        the admin of the same food bank, then it will return the admin's
        username and id as well as their email address.
      operationId: getFoodBankList
    post:
      tags:
        - food bank
      summary: Create food bank
      security:
        - BearerJwt: []
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodBank"
          description: The request was successful, and a new resource was created.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
          description: Authentication is required to access the requested resource. The
            client must include the appropriate credentials.
        "403":
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
          description: The server understood the request, but refuses to authorize it.
            Ensure the client has appropriate permissions.
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters: []
      description: |-
        must be an authenticated user to create a food bank
        (initially assume a 1-1 relationship between food bank and admin)
      operationId: createFoodBank
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - city
                - province
                - country
                - postal_code
                - staff_ids
                - phone
                - street
                - timezone
              properties:
                fax:
                  type: integer
                city:
                  type: string
                name:
                  type: string
                phone:
                  type: integer
                street:
                  type: string
                country:
                  type: string
                unit_no:
                  type: string
                website:
                  type: string
                province:
                  type: string
                staff_ids:
                  type: array
                  items:
                    type: integer
                charity_no:
                  type: string
                postal_code:
                  type: string
                timezone:
                  $ref: "#/components/schemas/timezone"
        required: true
        description: The food bank's static data. The authorized user will become the
          primary admin for this food bank.
...
components:
  schemas:
...
    User:
      type: object
      required:
        - username
        - email
        - id
      properties:
        id:
          type: integer
          minimum: 1
          examples:
            - 1
            - 2
            - 3
          description: the id of hte user
        email:
          type: string
          examples:
            - admin@gmail.com
          description: the email of the user
        username:
          type: string
          examples:
            - panther
            - cobra
          description: the username of the user
      description: A user is either a food bank admin or a food bank employee (note
        they are all considered staff)
...
    FoodBank:
      type: object
      required:
        - id
        - name
        - street
        - city
        - country
        - hours
        - admin
        - timzone
      properties:
        id:
          type: integer
        fax:
          type: integer
          description: the fax number for this food bank
        city:
          type: string
        name:
          type: string
          examples:
            - Fancy Food Bank
          description: the name of the food bank
        admin:
          $ref: "#/components/schemas/User"
        hours:
          type: string
        phone:
          type: integer
        staff:
          type: array
          items:
            type: integer
            minimum: 1
            description: the id of the employee
        street:
          type: string
          examples:
            - 555 Food Bank Way
            - 55 Pantry St
        country:
          type: string
        timzone:
          type: string
          examples:
            - America/Toronto
          description: this value is the IANA timezone value
        unit_no:
          type: string
          examples:
            - 121a
          description: Can be a number or a letter
        website:
          type: string
          examples:
            - info@website.com
          description: the food bank's website address
        latitude:
          type: number
          format: float
        province:
          type: string
        longitude:
          type: number
          format: float
        postal_code:
          type: string
        charity_registration_no:
          type: string
      description: The food bank or food pantry entity and all its details.
    timezone:
      type: string
      description: the IANA timezone string
      examples:
        - America/Toronto
  responses:
    Conflict:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
      description: The request could not be completed due to a conflict with the
        current state of the resource. Resolve the conflict and try again.
...
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
                description: the error message
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
    InternalServerErrror:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
  parameters:
...
    city:
      in: query
      name: city
      style: form
      schema:
        type: string
        examples:
          - toronto
          - metropolis
        description: the name of the city
      explode: true
      description: the name of a city
      allowReserved: false
      allowEmptyValue: false
...
    limit:
      in: query
      name: limit
      style: form
      schema:
        type: integer
      explode: true
      description: |-
        the number of items to return per page of results
        Default is 50
      allowReserved: false
      allowEmptyValue: false
...
    offset:
      in: query
      name: offset
      style: form
      schema:
        type: integer
      explode: true
      description: >-
        the number of rows to offset to (eg. if offset is 20, the first row
        retrieved will be row 20)

        Default is 0.
      allowReserved: false
      allowEmptyValue: false
    fb_name:
      in: query
      name: fb_name
      style: form
      schema:
        type: string
        examples:
          - Charity Food
          - Community Food Bank
        description: the name of the food bank
      explode: true
      description: the name of the food bank
      allowReserved: false
      allowEmptyValue: false
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
      bearerFormat: Json Web Token (JWT)
...